fix: Overflow on 32bit architectures when checking message length

+++ internal/gossiphttp/message.go
@@ -58,8 +58,8 @@ func readMessage(r io.Reader) ([]byte, error) {
 
 // writeMessage writes a message to an [io.Writer].
 func writeMessage(w io.Writer, message []byte) error {
-	if len(message) > math.MaxUint32 {
-		return fmt.Errorf("message length %d exceeds size limit %d", len(message), math.MaxUint32)
+	if uint(len(message)) > math.MaxUint32 {
+		return fmt.Errorf("message length %d exceeds size limit %d", len(message), uint32(math.MaxUint32))
 	}
 
 	header := headerPool.Get().(*header)
